Share query fixtures between linked-js and mem store
